SWATCH-1008: Refactor swatch-metrics jobs to use curl
