--- src/main/resources/application-metering-job.yaml
@@ -1,16 +1,0 @@
-METERING_TASK_TOPIC: ${clowder.kafka.topics[?(@.requestedName == 'platform.rhsm-subscriptions.metering-tasks')].name:platform.rhsm-subscriptions.metering-tasks}
-
-rhsm-subscriptions:
-  account-list-resource-location: ${ACCOUNT_LIST_RESOURCE_LOCATION:}
-  account-batch-size: ${ACCOUNT_BATCH_SIZE:1}
-
-  metering:
-    prometheus:
-      metric:
-        rangeInMinutes: ${OPENSHIFT_METERING_RANGE:60}
-        jobMaxAttempts: ${METERING_JOB_MAX_ATTEMPTS:3}
-        jobBackOffMaxInterval: ${METERING_JOB_BACK_OFF_MAX_INTERVAL:50000}
-        jobBackOffInitialInterval: ${METERING_JOB_BACK_OFF_INITIAL_INTERVAL:1000}
-        backOffMultiplier: ${METERING_JOB_BACK_OFF_MULTIPLIER:1.5}
-    tasks:
-      topic: ${METERING_TASK_TOPIC}
