+++ src/main/spec/internal-metering-api-spec.yaml
@@ -35,6 +35,19 @@ paths:
           $ref: "../../../spec/error-responses.yaml#/$defs/Forbidden"
         '500':
           $ref: "../../../spec/error-responses.yaml#/$defs/InternalServerError"
+  /internal/metering/sync:
+    description: 'Perform product metering for all accounts.'
+    put:
+      operationId: syncMetricsForAllAccounts
+      responses:
+        '200':
+          description: "Metering was successful."
+        '403':
+          $ref: "../../../spec/error-responses.yaml#/$defs/Forbidden"
+        '500':
+          $ref: "../../../spec/error-responses.yaml#/$defs/InternalServerError"
+      tags:
+        - internalProductMetering
   /internal/metering/{productTag}:
     description: 'Operations related to product metering.'
     post:
@@ -63,7 +76,7 @@ paths:
             type: integer
             minimum: 0
             description: "Defines the amount of time (in minutes) that will be used to calculate a metric query's start date based on a given end date.
-For example, given an end date of 2021-01-06T00:00:00Z and a rangeInMinutes of 60, the calculated start date will be: 2021-01-05T23:00:00Z.  Default (60) is store in application-metering-job.yaml"
+For example, given an end date of 2021-01-06T00:00:00Z and a rangeInMinutes of 60, the calculated start date will be: 2021-01-05T23:00:00Z.  Default (60) is store in application-openshift-metering-worker.yaml and application-metrics-rhel.yaml"
         - name: x-rh-swatch-synchronous-request
           in: header
           required: false
